first batch of rustup changes (still won't compile)

+++ src/waitstrategy.rs
@@ -6,10 +6,10 @@ use std::cmp::{min};
 pub trait WaitStrategy {
 
     /// Instantiate a new WaitStrategy. Must provide the size of the underlying buffer
-    fn new(ring_size: uint) -> Self;
+    fn new(ring_size: usize) -> Self;
 
     /// Get the underlying max buffer capacity
-    fn get_ring_size(&self) -> uint;
+    fn get_ring_size(&self) -> usize;
 
     /// Wait for the requested sequence, but return the largest available
     ///
@@ -27,7 +27,7 @@ pub trait WaitStrategy {
 /// This strategy should have the best perforamnce and keep caches hot, but will chew
 /// CPU while there is no work to be done.
 pub struct BusyWait {
-    ring_size: uint
+    ring_size: usize
 }
 
 impl BusyWait {
@@ -43,7 +43,7 @@ impl BusyWait {
                 return None;	// at same position as a dependency. we can't move
             }
             min_cursor = min(min_cursor, cursor);
-            debug!("					dep cursor: {}, ring_size: {}, sequence: {}", cursor, self.ring_size as int, sequence);
+            debug!("					dep cursor: {}, ring_size: {}, sequence: {}", cursor, self.ring_size as isize, sequence);
             debug!("					min_cursor: {}", min_cursor);
 
         }
@@ -52,13 +52,13 @@ impl BusyWait {
 }
 
 impl WaitStrategy for BusyWait {
-    fn new(ring_size: uint) -> BusyWait {
+    fn new(ring_size: usize) -> BusyWait {
         BusyWait {
             ring_size: ring_size
         }
     }
 
-    fn get_ring_size(&self) -> uint {
+    fn get_ring_size(&self) -> usize {
         self.ring_size
     }
